+++ admin/all.php
@@ -2,7 +2,5 @@
 require_once 'wcis-api.php';
 require_once 'wcis-checkout.php';
 
-require_once 'wcis-frontend.php';
 require_once 'wcis-ajax.php';
-
 require_once 'wcis-data.php';